Adds confirmation for delete actions on dashboard and detail view

+++ BoulderLogbook/Dashboard/Dashboard.swift
@@ -10,37 +10,51 @@ import ComposableArchitecture
 
 @Reducer
 struct Dashboard {
+    @Reducer(state: .equatable)
+    enum Destination {
+        case entryDetail(EntryDetail)
+        case confirmationDialog(ConfirmationDialogState<Confirmation>)
+
+        @CasePathable
+        enum Confirmation {
+            case delete
+        }
+    }
+
     @ObservableState
     struct State: Equatable {
-        @Presents var entryDetail: EntryDetail.State?
+        @Presents var destination: Destination.State?
         var sections: [Logbook.Section] = []
         var diagramPage = DiagramPage.State()
         var gradeSystems: [GradeSystem] = []
+        var entryToDelete: Logbook.Section.Entry.ID?
 
         var numberOfEntries: Int {
             sections.reduce(into: 0, { $0 += $1.entries.count})
         }
     }
     
-    enum Action: Equatable {
+    enum Action {
         case onAppear
         case fetchGradeSystems
         case receiveGradeSystems(TaskResult<[GradeSystem]>)
         case fetchSections
         case receiveSections(TaskResult<[Logbook.Section]>)
+        case showDeletionConfirmation
         case delete(Logbook.Section.Entry.ID)
         case deleteDidFinish(TaskResult<EntryClientResponse>)
         case edit(Logbook.Section.Entry)
         case setNavigation(Logbook.Section.Entry)
-        case entryDetail(PresentationAction<EntryDetail.Action>)
         case diagramPage(DiagramPage.Action)
+        case destination(PresentationAction<Destination.Action>)
 
         enum EntryClientResponse { case finished }
     }
     @Dependency(LogbookEntryClient.self) var logbookEntryClient
     @Dependency(GradeSystemClient.self) var gradeSystemClient
-    
-    var body: some Reducer<State, Action> {
+    @Dependency(\.dismiss) var dismiss
+
+    var body: some ReducerOf<Self> {
         Scope(state: \.diagramPage, action: /Action.diagramPage) {
             DiagramPage()
         }
@@ -84,14 +98,37 @@ struct Dashboard {
                 state.sections = sections.sorted(
                     by: { $0.date > $1.date }
                 )
-                
+
             case let .delete(id),
-                 let .entryDetail(.presented(.delete(id))):
+                 let .destination(.presented(.entryDetail(.delete(id)))) :
+                state.entryToDelete = id
+                return .run { send in
+                    try await Task.sleep(for: .milliseconds(100))
+                    await send(.showDeletionConfirmation)
+                }
+
+            case .showDeletionConfirmation:
+                state.destination = .confirmationDialog(
+                    ConfirmationDialogState {
+                        TextState("Warning")
+                    } actions: {
+                        ButtonState(role: .destructive, action: .delete) {
+                            TextState("Delete")
+                        }
+                    } message: {
+                        TextState("Are you sure you want to delete this entry?")
+                    }
+                )
+
+            case .destination(.presented(.confirmationDialog(.delete))):
+                guard let entryToDelete = state.entryToDelete else {
+                    return .none
+                }
                 return .run { send in
                     await send(
                         .deleteDidFinish(
                             TaskResult {
-                                await logbookEntryClient.deleteEntry(id)
+                                await logbookEntryClient.deleteEntry(entryToDelete)
                                 return .finished
                             }
                         )
@@ -99,17 +136,19 @@ struct Dashboard {
                 }
 
             case .deleteDidFinish:
-                state.entryDetail = nil
+                state.destination = nil
                 return .merge(
                     .send(.fetchSections),
                     .send(.diagramPage(.fetchEntries))
                 )
-                
+
             case let .setNavigation(entry):
                 if let system = state.gradeSystems.first(where: { $0.id == entry.gradeSystem }) {
-                    state.entryDetail = EntryDetail.State(
-                        entry: entry,
-                        gradeSystem: system
+                    state.destination = .entryDetail(
+                        EntryDetail.State(
+                            entry: entry,
+                            gradeSystem: system
+                        )
                     )
                 }
 
@@ -117,8 +156,6 @@ struct Dashboard {
             }
             return .none
         }
-        .ifLet(\.$entryDetail, action: \.entryDetail) {
-            EntryDetail()
-        }
+        .ifLet(\.$destination, action: \.destination)
     }
 }
